requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic[email]==2.11.3
python-multipart==0.0.20
jinja2==3.1.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==1.1.2
weasyprint==65.1
aiofiles==24.1.0
